# app/openapi.yml
openapi: 3.0.3
info:
  title: SISA koppel api
  description: Dit is het koppelstuk tussen Mijn Amsterdam en SISA
  version: 0.0.1
paths:
  /api/subsidie/summary:
    parameters:
      - name: x-saml-attribute-token1
        in: header
        description: TMA encoded BSN
        required: true
        schema:
          type: string
    get:
      summary: Informatie m.b.t. subsidies op basis van BSN
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    $ref: '#/components/schemas/UserInfo'
                  status:
                    type: string
                    enum:
                      - 'OK'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /status/health:
    get:
      summary: returns "OK"
      responses:
        '200':
          description: Returns "OK"
          content:
            application/json:
              schema:
                type: string
                enum:
                  - 'OK'
components:
  schemas:
    UserInfo:
      type: object
      properties:
        isKnown:
          type: boolean
        notifications:
          type: array
          items:
            $ref: "#/components/schemas/MyNotification"
    MyNotification:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        datePublished:
          $ref: "#/components/schemas/DateTime"
          nullable: false
    Date:
      type: string
      format: date
      example: '2020-06-16'
    DateTime:
      type: string
      format: date-time
      example: '2020-06-16T:13:40:00'
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          enum:
            - 'ERROR'
        message:
          type: string